main: refactor main package - Make argumentList a smarter object - Wrap most of the main logic in a run function which returns an error - Make logger global (for main package only) - Minor improvements to error messages and formatting
