+++ src/kubernetes.go
@@ -1,17 +1,12 @@
 package main
 
 import (
-	"errors"
 	"fmt"
 	"os"
-	"path"
-	"strings"
 	"time"
 
-	sdkArgs "github.com/newrelic/infra-integrations-sdk/args"
 	"github.com/newrelic/infra-integrations-sdk/integration"
 	"github.com/newrelic/infra-integrations-sdk/log"
-	"github.com/sirupsen/logrus"
 
 	"github.com/newrelic/nri-kubernetes/v2/src/apiserver"
 	"github.com/newrelic/nri-kubernetes/v2/src/client"
@@ -20,41 +15,15 @@ import (
 	"github.com/newrelic/nri-kubernetes/v2/src/data"
 	"github.com/newrelic/nri-kubernetes/v2/src/featureflag"
 	"github.com/newrelic/nri-kubernetes/v2/src/ksm"
-	clientKsm "github.com/newrelic/nri-kubernetes/v2/src/ksm/client"
+	ksmclient "github.com/newrelic/nri-kubernetes/v2/src/ksm/client"
 	"github.com/newrelic/nri-kubernetes/v2/src/kubelet"
-	clientKubelet "github.com/newrelic/nri-kubernetes/v2/src/kubelet/client"
-	metric2 "github.com/newrelic/nri-kubernetes/v2/src/kubelet/metric"
+	kubeletclient "github.com/newrelic/nri-kubernetes/v2/src/kubelet/client"
+	kubeletmetric "github.com/newrelic/nri-kubernetes/v2/src/kubelet/metric"
 	"github.com/newrelic/nri-kubernetes/v2/src/metric"
 	"github.com/newrelic/nri-kubernetes/v2/src/network"
 	"github.com/newrelic/nri-kubernetes/v2/src/scrape"
 	"github.com/newrelic/nri-kubernetes/v2/src/storage"
 )
-
-type argumentList struct {
-	sdkArgs.DefaultArgumentList
-	Timeout                      int    `default:"5000" help:"timeout in milliseconds for calling metrics sources"`
-	ClusterName                  string `help:"Identifier of your cluster. You could use it later to filter data in your New Relic account"`
-	DiscoveryCacheDir            string `default:"/var/cache/nr-kubernetes" help:"The location of the cached values for discovered endpoints. Obsolete, use CacheDir instead."`
-	CacheDir                     string `default:"/var/cache/nr-kubernetes" help:"The location where to store various cached data."`
-	DiscoveryCacheTTL            string `default:"1h" help:"Duration since the discovered endpoints are stored in the cache until they expire. Valid time units: 'ns', 'us', 'ms', 's', 'm', 'h'"`
-	APIServerCacheTTL            string `default:"5m" help:"Duration to cache responses from the API Server. Valid time units: 'ns', 'us', 'ms', 's', 'm', 'h'. Set to 0s to disable"`
-	APIServerCacheK8SVersionTTL  string `default:"3h" help:"Duration to cache the kubernetes version responses from the API Server. Valid time units: 'ns', 'us', 'ms', 's', 'm', 'h'. Set to 0s to disable"`
-	EtcdTLSSecretName            string `help:"Name of the secret that stores your ETCD TLS configuration"`
-	EtcdTLSSecretNamespace       string `default:"default" help:"Namespace in which the ETCD TLS secret lives"`
-	DisableKubeStateMetrics      bool   `default:"false" help:"Used to disable KSM data fetching. Defaults to 'false''"`
-	KubeStateMetricsURL          string `help:"kube-state-metrics URL. If it is not provided, it will be discovered."`
-	KubeStateMetricsPodLabel     string `help:"discover KSM using Kubernetes Labels."`
-	KubeStateMetricsPort         int    `default:"8080" help:"port to query the KSM pod. Only works together with the pod label discovery"`
-	KubeStateMetricsScheme       string `default:"http" help:"scheme to query the KSM pod ('http' or 'https'). Only works together with the pod label discovery"`
-	KubeStateMetricsNamespace    string `default:"" help:"namespace to query the KSM pod. By default, all namespaces will be queried"`
-	DistributedKubeStateMetrics  bool   `default:"false" help:"Set to enable distributed KSM discovery. Requires that KubeStateMetricsPodLabel is set. Disabled by default."`
-	APIServerSecurePort          string `default:"" help:"Set to query the API Server over a secure port. Disabled by default"`
-	SchedulerEndpointURL         string `help:"Set a custom endpoint URL for the kube-scheduler endpoint."`
-	EtcdEndpointURL              string `help:"Set a custom endpoint URL for the Etcd endpoint."`
-	ControllerManagerEndpointURL string `help:"Set a custom endpoint URL for the kube-controller-manager endpoint."`
-	APIServerEndpointURL         string `help:"Set a custom endpoint URL for the API server endpoint."`
-	NetworkRouteFile             string `help:"Route file to get the default interface from. If left empty on Linux /proc/net/route will be used by default"`
-}
 
 const (
 	// we use '/var/cache/nr-kubernetes' as the temp cache dir rather than
@@ -78,19 +47,6 @@ var (
 	integrationVersion    = "dev"
 	integrationCommitHash = "unknown"
 )
-
-var args argumentList
-
-func getCacheDir(subDirectory string) string {
-	cacheDir := args.CacheDir
-
-	// accept the old cache directory argument if it's explicitly set
-	if args.DiscoveryCacheDir != defaultCacheDir {
-		cacheDir = args.DiscoveryCacheDir
-	}
-
-	return path.Join(cacheDir, subDirectory)
-}
 
 func controlPlaneJobs(
 	logger log.Logger,
@@ -183,44 +139,31 @@ func controlPlaneJobs(
 	return jobs, nil
 }
 
-func main() {
-	integration, err := integration.New(integrationName, integrationVersion, integration.Args(&args))
-	var jobs []*scrape.Job
-	exitLog := fmt.Sprintf("Integration %q exited", integrationName)
-	if err != nil {
-		defer log.Debug(exitLog)
-		log.Fatal(err) // Global logs used as args processed inside NewIntegrationProtocol2
-	}
+func run() error {
+	args := &argumentList{}
+
+	intgr, err := integration.New(integrationName, integrationVersion, integration.Args(args))
 
 	logger := log.NewStdErr(args.Verbose)
-	defer func() {
-		if r := recover(); r != nil {
-			recErr, ok := r.(*logrus.Entry)
-			if ok {
-				recErr.Fatal(recErr.Message)
-			} else {
-				panic(r)
-			}
-		}
-	}()
 
-	defer logger.Debugf(exitLog)
+	var jobs []*scrape.Job
+	if err != nil {
+		return fmt.Errorf("creating integration: %w", err)
+	}
+
 	logger.Debugf("Integration %q ver. %s (git %s) started", integrationName, integrationVersion, integrationCommitHash)
 	if args.ClusterName == "" {
-		logger.Errorf("cluster_name argument is mandatory")
-		os.Exit(1)
+		return fmt.Errorf("cluster_name argument is mandatory")
 	}
 
 	nodeName := os.Getenv(nodeNameEnvVar)
 	if nodeName == "" {
-		logger.Errorf("%s env var should be provided by Kubernetes and is mandatory", nodeNameEnvVar)
-		os.Exit(1)
+		return fmt.Errorf("could not find %q in env, which should have been set by Kubernetes", nodeNameEnvVar)
 	}
 
-	// TODO: Map to v3
-	//if !args.All && !args.Metrics {
-	//	os.Exit(1)
-	//}
+	if !args.HasMetrics() {
+		return fmt.Errorf("integration only supports the -metrics or -all commands")
+	}
 
 	ttl, err := time.ParseDuration(args.DiscoveryCacheTTL)
 	if err != nil {
@@ -230,72 +173,73 @@ func main() {
 
 	timeout := time.Millisecond * time.Duration(args.Timeout)
 
-	innerKubeletDiscoverer, err := clientKubelet.NewDiscoverer(nodeName, logger)
+	innerKubeletDiscoverer, err := kubeletclient.NewDiscoverer(nodeName, logger)
 	if err != nil {
-		logger.Errorf("Error during Kubelet auto discovering process: %v", err)
+		return fmt.Errorf("building kubelet discoverer: %w", err)
 	}
-	cacheStorage := storage.NewJSONDiskStorage(getCacheDir(discoveryCacheDir))
 
-	defaultNetworkInterface, err := network.CachedDefaultInterface(
-		logger, args.NetworkRouteFile, cacheStorage, ttl)
+	cacheStorage := storage.NewJSONDiskStorage(args.cacheDir(discoveryCacheDir))
+
+	defaultNetworkInterface, err := network.CachedDefaultInterface(logger, args.NetworkRouteFile, cacheStorage, ttl)
 	if err != nil {
-		logger.Warnf("Error finding default network interface: %v", err)
+		logger.Warnf("Could not find default network interface: %v", err)
 	}
-	kubeletDiscoverer := clientKubelet.NewDiscoveryCacher(innerKubeletDiscoverer, cacheStorage, ttl, logger)
+
+	kubeletDiscoverer := kubeletclient.NewDiscoveryCacher(innerKubeletDiscoverer, cacheStorage, ttl, logger)
 
 	kubeletClient, err := kubeletDiscoverer.Discover(timeout)
 	if err != nil {
-		logger.Errorf("Error discovering kubelet: %v", err)
-		os.Exit(1)
+		return fmt.Errorf("discovering kubelet: %w", err)
 	}
+
 	kubeletNodeIP := kubeletClient.NodeIP()
 	logger.Debugf("Kubelet node IP = %s", kubeletNodeIP)
 
 	k8s, err := client.NewKubernetes(false)
 	if err != nil {
-		logger.Errorf("Error building kubernetes client: %v", err)
-		os.Exit(1)
+		return fmt.Errorf("building kubernetes client: %w", err)
 	}
 
 	if !args.DisableKubeStateMetrics {
 		var ksmClients []client.HTTPClient
 		var ksmNodeIP string
 		if args.DistributedKubeStateMetrics {
-			ksmDiscoverer, err := getMultiKSMDiscoverer(kubeletNodeIP, logger)
+			ksmDiscoverer, err := args.multiKSMDiscoverer(logger, kubeletNodeIP)
 			if err != nil {
-				logger.Errorf("Error getting multiKSM discoverer: %v", err)
-				os.Exit(1)
+				return fmt.Errorf("getting multiKSM discoverer: %v", err)
 			}
-			ksmDiscoveryCache := clientKsm.NewDistributedDiscoveryCacher(ksmDiscoverer, cacheStorage, ttl, logger)
+
+			ksmDiscoveryCache := ksmclient.NewDistributedDiscoveryCacher(ksmDiscoverer, cacheStorage, ttl, logger)
 			ksmClients, err = ksmDiscoveryCache.Discover(timeout)
+			if err != nil {
+				return fmt.Errorf("discovering KSM: %w", err)
+			}
+
 			logger.Debugf("found %d KSM clients:", len(ksmClients))
 			for _, c := range ksmClients {
 				logger.Debugf("- node IP: %s", c.NodeIP())
 			}
-			if err != nil {
-				logger.Errorf("Error discovering KSM: %v", err)
-				os.Exit(1)
-			}
 			ksmNodeIP = kubeletNodeIP
 		} else {
-			innerKSMDiscoverer, err := getKSMDiscoverer(logger)
+			innerKSMDiscoverer, err := args.ksmDiscoverer(logger)
 			if err != nil {
-				logger.Errorf("Error getting KSM discoverer: %v", err)
-				os.Exit(1)
+				return fmt.Errorf("getting KSM discoverer: %v", err)
 			}
-			ksmDiscoverer := clientKsm.NewDiscoveryCacher(innerKSMDiscoverer, cacheStorage, ttl, logger)
+
+			ksmDiscoverer := ksmclient.NewDiscoveryCacher(innerKSMDiscoverer, cacheStorage, ttl, logger)
 			ksmClient, err := ksmDiscoverer.Discover(timeout)
 			if err != nil {
-				logger.Errorf("Error discovering KSM: %v", err)
-				os.Exit(1)
+				return fmt.Errorf("discovering standalone KSM: %v", err)
 			}
+
 			ksmNodeIP = ksmClient.NodeIP()
 			// we only scrape KSM when we are on the same Node as KSM
 			if kubeletNodeIP == ksmNodeIP {
 				ksmClients = append(ksmClients, ksmClient)
 			}
 		}
-		logger.Debugf("KSM Node = %s", ksmNodeIP)
+		logger.Debugf("Found KSM in node %q", ksmNodeIP)
+
 		for _, ksmClient := range ksmClients {
 			ksmGrouper := ksm.NewGrouper(ksmClient, metric.KSMQueries, logger, k8s)
 			jobs = append(jobs, scrape.NewScrapeJob("kube-state-metrics", ksmGrouper, metric.KSMSpecs))
@@ -305,6 +249,7 @@ func main() {
 	apiServerClient := apiserver.NewClient(k8s)
 
 	ttlAPIServerCacheK8SVersion, err := time.ParseDuration(args.APIServerCacheK8SVersionTTL)
+
 	if err != nil {
 		logger.Errorf(
 			"Error while parsing the api server cache TTL value for the kubernetes server version, defaulting to %s: %v",
@@ -317,16 +262,18 @@ func main() {
 	if ttlAPIServerCacheK8SVersion != time.Duration(0) {
 		apiServerClientK8sVersion = apiserver.NewFileCacheClientWrapper(
 			apiServerClient,
-			getCacheDir(apiserverCacheDirK8sVersion),
+			args.cacheDir(apiserverCacheDirK8sVersion),
 			ttlAPIServerCacheK8SVersion,
 		)
 	} else {
 		apiServerClientK8sVersion = apiServerClient
 	}
+
 	k8sVersion, err := apiServerClientK8sVersion.GetServerVersion()
 	if err != nil {
 		logger.Errorf("Error getting the kubernetes server version: %v", err)
 	}
+
 	enableStaticPodsStatus := featureflag.StaticPodsStatus(k8sVersion)
 
 	ttlAPIServerCache, err := time.ParseDuration(args.APIServerCacheTTL)
@@ -337,11 +284,11 @@ func main() {
 
 	if ttlAPIServerCache != time.Duration(0) {
 		apiServerClient = apiserver.NewFileCacheClientWrapper(apiServerClient,
-			getCacheDir(apiserverCacheDir),
+			args.cacheDir(apiserverCacheDir),
 			ttlAPIServerCache)
 	}
 
-	podsFetcher := metric2.NewPodsFetcher(logger, kubeletClient, enableStaticPodsStatus).FetchFuncWithCache()
+	podsFetcher := kubeletmetric.NewPodsFetcher(logger, kubeletClient, enableStaticPodsStatus).FetchFuncWithCache()
 	cpJobs, err := controlPlaneJobs(
 		logger,
 		apiServerClient,
@@ -360,7 +307,7 @@ func main() {
 	)
 
 	if err != nil {
-		logger.Errorf("couldn't configure control plane components jobs: %v", err)
+		logger.Errorf("Couldn't configure control plane components jobs, skipping control plane monitoring: %v", err)
 	} else {
 		jobs = append(jobs, cpJobs...)
 	}
@@ -372,7 +319,7 @@ func main() {
 		apiServerClient,
 		defaultNetworkInterface,
 		podsFetcher,
-		metric2.CadvisorFetchFunc(kubeletClient, metric.CadvisorQueries),
+		kubeletmetric.CadvisorFetchFunc(kubeletClient, metric.CadvisorQueries),
 	)
 	jobs = append(jobs, scrape.NewScrapeJob("kubelet", kubeletGrouper, metric.KubeletSpecs))
 
@@ -380,7 +327,7 @@ func main() {
 	for _, job := range jobs {
 		logger.Debugf("Running job: %s", job.Name)
 		start := time.Now()
-		result := job.Populate(integration, args.ClusterName, logger, k8sVersion)
+		result := job.Populate(intgr, args.ClusterName, logger, k8sVersion)
 		measured := time.Since(start)
 		logger.Debugf("Job %s took %s", job.Name, measured.Round(time.Millisecond))
 
@@ -394,89 +341,12 @@ func main() {
 	}
 
 	if successfulJobs == 0 {
-		logger.Errorf("No data was populated")
-		os.Exit(1)
+		return fmt.Errorf("getting k8s data, all jobs failed")
 	}
 
-	if err := integration.Publish(); err != nil {
-		logger.Errorf("Error rendering integration output: %v", err)
-		os.Exit(1)
-	}
-}
-
-func getKSMDiscoverer(logger log.Logger) (client.Discoverer, error) {
-	k8sClient, err := client.NewKubernetes( /* tryLocalKubeconfig */ false)
-	if err != nil {
-		return nil, fmt.Errorf("initializing Kubernetes client: %w", err)
+	if err := intgr.Publish(); err != nil {
+		return fmt.Errorf("rendering integration output: %w", err)
 	}
 
-	config := clientKsm.DiscovererConfig{
-		K8sClient: k8sClient,
-		Logger:    logger,
-		Namespace: args.KubeStateMetricsNamespace,
-	}
-
-	// It's important this one is before the NodeLabel selector, for backwards compatibility.
-	if args.KubeStateMetricsURL != "" {
-		// Remove /metrics suffix if present
-		args.KubeStateMetricsURL = strings.TrimSuffix(args.KubeStateMetricsURL, "/metrics")
-
-		logger.Debugf("Discovering KSM using static endpoint (KUBE_STATE_METRICS_URL=%s)", args.KubeStateMetricsURL)
-
-		config.OverridenEndpoint = args.KubeStateMetricsURL
-
-		return clientKsm.NewDiscoverer(config)
-	}
-
-	if args.KubeStateMetricsPodLabel != "" {
-		logger.Debugf("Discovering KSM using Pod Label (KUBE_STATE_METRICS_POD_LABEL)")
-
-		config := clientKsm.PodLabelDiscovererConfig{
-			KSMPodLabel:  args.KubeStateMetricsPodLabel,
-			KSMPodPort:   args.KubeStateMetricsPort,
-			KSMScheme:    args.KubeStateMetricsScheme,
-			KSMNamespace: args.KubeStateMetricsNamespace,
-			Logger:       logger,
-			K8sClient:    k8sClient,
-		}
-
-		discoverer, err := clientKsm.NewPodLabelDiscoverer(config)
-		if err != nil {
-			return nil, fmt.Errorf("creating KSM pod label discoverer: %w", err)
-		}
-
-		return discoverer, nil
-	}
-
-	logger.Debugf("Discovering KSM using DNS / k8s ApiServer (default)")
-
-	return clientKsm.NewDiscoverer(config)
-}
-
-func getMultiKSMDiscoverer(nodeIP string, logger log.Logger) (client.MultiDiscoverer, error) {
-	k8sClient, err := client.NewKubernetes( /* tryLocalKubeconfig */ false)
-	if err != nil {
-		return nil, fmt.Errorf("initializing Kubernetes client: %w", err)
-	}
-
-	if args.KubeStateMetricsPodLabel == "" {
-		return nil, errors.New("multi KSM discovery set without a KUBE_STATE_METRICS_POD_LABEL")
-	}
-
-	logger.Debugf("Discovering distributed KSMs using pod labels from KUBE_STATE_METRICS_POD_LABEL")
-
-	config := clientKsm.DistributedPodLabelDiscovererConfig{
-		KSMPodLabel:  args.KubeStateMetricsPodLabel,
-		KSMNamespace: args.KubeStateMetricsNamespace,
-		NodeIP:       nodeIP,
-		K8sClient:    k8sClient,
-		Logger:       logger,
-	}
-
-	client, err := clientKsm.NewDistributedPodLabelDiscoverer(config)
-	if err != nil {
-		return nil, fmt.Errorf("creating new distributed pod label discoverer: %w", err)
-	}
-
-	return client, nil
+	return nil
 }
