+++ src/kubernetes_test.go
@@ -2,10 +2,10 @@ package main
 
 import (
 	"fmt"
+	"github.com/sirupsen/logrus"
 	"testing"
 	"time"
 
-	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
 
 	"github.com/newrelic/nri-kubernetes/v2/src/apiserver"
@@ -13,9 +13,9 @@ import (
 	"github.com/newrelic/nri-kubernetes/v2/src/definition"
 )
 
-var logger = logrus.StandardLogger()
-
 func TestControlPlaneJobs(t *testing.T) {
+	logger := logrus.StandardLogger()
+
 	nodeName := "ip-10.0.2.15"
 	nodeIP := "10.0.2.15"
 	// Setup the data returned when querying the kubelet for the pods
